rename file reducer to upload reducer for clarity
